Reduce inode deletion cluster lock hold times
A lot of file extent manipulation cluster lock hold times are
proportional to the number of extents being manipulated because they
were modeled on i_mutex patterns.

This can be a problem when there are a truly immense number of extents
to work with.  Other callers trying to acquire the inode cluster lock
will block and eventually cause hunt task timeout messages.

We have a test that specifically triggers this all the time.  It's made
a worse when running in debugging kernels or loaded infrastructure that
causes everything to slow down.

The typical solution for this pattern is to limit the amount of work
done under each lock hold.  That's hard to solve universally in this
case because most callers working on live inodes have visible atomicity
promises that stop us from making partial progress visible to other lock
acquisition attempts.

So this change is a specific attempt to stop the problem case of
deleting the extent items once an inode is fully unlinked and no longer
referenced.  In this case we can back out of deletion if we spend too
many transactions deleting extent items.  The orphan scanners were
already built to support picking up after this kind of interrupted
partial progress.

Signed-off-by: Zach Brown <zab@versity.com>