Add t_ino_has_items test helper
Add a test helper which tests if an inode number has existing fs items.
This used to be a private function used by orphan inode deletion testing
but we want to use it in other places so we'll promote it up to be
available for all tests, with a new name that more precisely describes
what it does.

Signed-off-by: Zach Brown <zab@versity.com>