Fix wrapping detection code blocks.
A new test from xfstests (generic/525) exposes that we poorly detect
wrapping of u64/s64 values in a few cases. The key test is that we
should always be able to fully use the last block, where (offset +
length) is larger than (e.g.) LLONG_MAX. In other words, we want to make
sure length is not negative, and (offset + (length - 1)) doesn't wrap.

scoutfs_client_alloc_inodes() checks that u64 inode numbers do not wrap.
We allocate these in 10*1024 hunks, and in practice we will never hit
the issue since (2^64 % 10240) is not 0.

scoutfs_fallocate() may hit the wraping issue.

scoutfs_data_wait_check() is the case that was exposed by generic/525.

Our own test `simple-release-extents` almost exposed the wrapping issue
in scoutfs_ioc_release(). We adjust the test to test the edge case and
the wrapping case properly.

Signed-off-by: Auke Kok <auke.kok@versity.com>