Do not iterate on iter->iov when the iter is not an iovec.
The kernel's splice implementation uses an ITER_PIPE type iter_iov
which means we shouldn't try and iterate over iter->iov - there's
just nothing there. We can just perform checks on the inode/pos/count
and be done here.

The check for !iter_type_iov() here catches pipe and all other non-iov
types.

This issue isn't exposed in el8 because the kernel doesn't yet fall
back to this code path. But it will hit the same error condition in
el9 and the code will then return -EINVAL when it tries to sendfile().

Signed-off-by: Auke Kok <auke.kok@versity.com>